feat: add wildcard mode UI to settings panel
